+++ src/test/java/mainTest.java
@@ -71,18 +71,6 @@ class mainTest {
         assertEquals(actual, expected);
     }
 
-
-    void kiumsellpass() {
-        String stockCode = "T02";
-        int count = 1;
-        int price = 2000;
-
-        KiumTradingSystem app = new KiumTradingSystem();
-        String actual = app.buy(stockCode,count,price);
-        String expected = String.format("%s를 %d 가격에 매도하였음", stockCode,price);
-        assertEquals(actual, expected);
-    }
-
     void namusellypass() {
         String stockCode = "T02";
         int count = 1;
@@ -93,5 +81,4 @@ class mainTest {
         String expected = String.format("%s를 %d 가격에 매도하였음", stockCode,price);
         assertEquals(actual, expected);
     }
-
 }
